Bug report script now runs faster.
Find runs at a max-depth of 3, instead of 4,  as
i've never found a distro that embeds xscreensaver
directories quite that far.

It now also searchs for directories starting with
xscreensaver, so that xscreensaver-misc and others
can be found.

Simple changes, but find is the main bottleneck of
the script, so it should run magnitudes faster now.