Reformatted font/image data header files.
I wrote a script to reformat the data. There's now 20 numbers a line, meaning
no line exceeds 80 characters, and the file in general is more compact.

I also updated the comments in matrix2.h that denoted the start of a new image.
The offsets were off by a few characters.

Finally, I added instructions for making and inserting a new image into matrix2.h.

The net result is a much cleaner header that's easier to maintain.